+++ test/a/module_b.js
@@ -1,0 +1,3 @@
+var c = require('module_c');
+
+exports.result = "loaded "+c.hello;