+++ test/test-require.js
@@ -14,11 +14,17 @@ describe("CommonJS module", function() {
     });
 });
 
-
 var ex = require('./requiredexample');
 var m = require('./a/b');
 
 describe("require function", function() {
+
+    it("should be able to load modules from the main script directory without relative path", function(){
+        var a = require('module_a');
+        expect(a.result).toEqual("okay loaded world");
+    });
+
+
     it("should import exported variable of requiredexample module", function(){
         expect('ex' in slimerEnv).toBeTruthy();
         expect(ex.myExample).toEqual("foo");
@@ -29,8 +35,30 @@ describe("require function", function() {
         expect(m.identity.firstName).toEqual("Laurent");
     });
 
-    it("has a modification in paths and can read a module in an other path", function(){
+    it("has a modification in paths and can read a module in an other absolute path", function(){
         var path = fs.absolute(phantom.libraryPath + '/../test-modules/');
+        require.paths.push (path);
+        var b = require('b');
+        expect(b.bIsLoaded).toEqual('yes');
+        expect(b.cIsLoaded).toEqual('yes');
+        expect(b.idName).toEqual('bob');
+        var e= require('d/e');
+        expect(e.readedFromF).toEqual('var of f');
+        // remove the path now...
+        require.paths.pop();
+        var ok = false;
+        try {
+            var b = require('b');
+        }
+        catch(e) {
+            // the b module should not be found
+            ok = true;
+        }
+        expect(ok).toBeTruthy()
+    });
+
+    it("has a modification in paths and can read a module in an other relative path", function(){
+        var path = '../test-modules/';
         require.paths.push (path);
         var b = require('b');
         expect(b.bIsLoaded).toEqual('yes');
@@ -57,7 +85,6 @@ describe("require function", function() {
         expect(access.accessToGlobalFunction).toEqual("result from a global function")
     })
 });
-
 
 describe("the loaded module requiredexample", function() {