+++ docs/release-notes.rst
@@ -17,6 +17,10 @@ Improvements
 - Improve performance of ``webpage.render()``
 - ``--debug=true`` now displays arguments of the script and SlimerJS's configuration
 
+Fixed bugs
+----------
+- ``require.paths`` should accept relative paths (issue #73)
+
 version 0.8.2
 =============